+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Current
  
 - Change `launch_anvil()` to use latest hardfork by default instead of `london`
+- Add: `wait_and_broadcast_multiple_nodes_mev_blocker()` for [MEV Blocker](https://mevblocker.io) - because the tx
+  broadcast must be sequential
 
 # 0.26.1